+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/PodcastManagerImpl.kt
@@ -6,6 +6,7 @@ import au.com.shiftyjelly.pocketcasts.models.db.helper.TopPodcast
 import au.com.shiftyjelly.pocketcasts.models.entity.Folder
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
+import au.com.shiftyjelly.pocketcasts.models.entity.TrendingPodcast
 import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
 import au.com.shiftyjelly.pocketcasts.models.to.AutoArchiveAfterPlaying
 import au.com.shiftyjelly.pocketcasts.models.to.AutoArchiveInactive
@@ -582,6 +583,10 @@ class PodcastManagerImpl @Inject constructor(
 
     override fun addFolderPodcast(podcast: Podcast) {
         podcastDao.insert(podcast)
+    }
+
+    override suspend fun replaceTrendingPodcasts(podcasts: List<TrendingPodcast>) {
+        podcastDao.replaceAllTrendingPodcasts(podcasts)
     }
 
     override fun updatePodcast(podcast: Podcast) {