+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/shownotes/ShowNotesProcessor.kt
@@ -74,5 +74,6 @@ class ShowNotesProcessor @Inject constructor(
         title = title,
         imageUrl = image,
         url = url,
+        isEmbedded = false,
     )
 }